requirements.txt
flask>=2.3.0
numpy>=1.21.0
torch>=2.0.0
PyPDF2>=3.0.0
markdown>=3.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
transformers>=4.30.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
accelerate>=0.20.0
